refactor: 가격 long -> int 수정 - #3

+++ src/main/java/megabooks/megabooks/domain/book/entity/Book.java
@@ -15,7 +15,7 @@ public class Book extends BaseEntity {
     private String bookTitle;
     private String bookAuthor;
     private String bookPublisher;
-    private long bookPrice;
+    private int bookPrice;
     private String bookImg;
 
     @Enumerated(value = EnumType.STRING)
